test(tc): keep annotation rendering in test support
